move ClientHintsInterceptor into MangaFire
