+++ core/src/main/kotlin/keiyoushi/source/KeiSource.kt
@@ -12,7 +12,6 @@ import eu.kanade.tachiyomi.source.model.SManga
 import eu.kanade.tachiyomi.source.model.SMangaUpdate
 import eu.kanade.tachiyomi.source.online.HttpSource
 import keiyoushi.network.CacheControlInterceptor
-import keiyoushi.network.ClientHintsInterceptor
 import keiyoushi.network.RateLimitInterceptor
 import keiyoushi.utils.applicationContext
 import keiyoushi.utils.firstInstanceOrNull
@@ -113,9 +112,6 @@ abstract class KeiSource : HttpSource() {
                     add(cloudflareInterceptor)
                 }
             }
-
-            // automatically set client hints based on user agent
-            addInterceptor(ClientHintsInterceptor)
 
             // last application interceptor
             addInterceptor(CompressionInterceptor(Brotli, Gzip, Zstd))
